+++ src/generic-table/generic-table.component.ts
@@ -444,9 +444,12 @@ export class GenericTableComponent implements OnInit, OnChanges {
   // TODO: move to helper functions
   /** Sort by column order */
   private getColumnOrder = function(a,b) {
+    if (a.columnOrder === undefined) {
+      return -1;
+    }
     if (a.columnOrder < b.columnOrder)
       return -1;
-    if (a.columnOrder > b.columnOrder || typeof a.columnOrder === 'undefined')
+    if (a.columnOrder > b.columnOrder)
       return 1;
     return 0;
   };